9.4.1

+++ Italbytz.Ports.Algorithms.EA/Italbytz.Ports.Algorithms.EA/SearchSpace/IMonomial.cs
@@ -19,7 +19,7 @@ public interface IMonomial<TCategory> : ICloneable
     List<ILiteral<TCategory>> Literals { get; set; }
     float[] Weights { get; set; }
     
-    public double[] Evaluate(TCategory[] input);
+    public float[] Evaluate(TCategory[] input);
     
     public float[][] Predictions { get; set; }
     int Size { get; }
